+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+* 'removed' issue component state
+
 ## [0.10.4] - 2024-10-18
 
 ## Fixed